doc-source/requirements.txt
git+https://github.com/sphinx-toolbox/sphinx-toolbox-experimental
git+https://github.com/domdfcoding/emojicodes@close-files
autodocsumm>=0.2.0
default-values>=0.4.2
domdf-sphinx-theme>=0.3.0
extras-require>=0.2.0
seed-intersphinx-mapping>=0.3.1
sphinx>=3.0.3
sphinx-autofixture>=0.2.1
sphinx-copybutton>=0.2.12
sphinx-debuginfo>=0.1.0
sphinx-highlights>=0.4.0
sphinx-notfound-page>=0.5
sphinx-prompt>=1.1.0
sphinx-pyproject>=0.1.0
sphinx-tabs>=1.1.13
sphinx-toolbox>=2.10.0
sphinxcontrib-httpdomain>=1.7.0
sphinxemoji>=0.1.6
toctree-plus>=0.1.0
